% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrapPlot}
\alias{bootstrapPlot}
\alias{bootstrapPlot.nlmixr2FitCore}
\title{Produce delta objective function for bootstrap}
\usage{
bootstrapPlot(x, ...)

\method{bootstrapPlot}{nlmixr2FitCore}(x, ...)
}
\arguments{
\item{x}{fit object}

\item{...}{other parameters}
}
\value{
Fit traceplot or nothing.
}
\description{
Produce delta objective function for bootstrap
}
\references{
R Niebecker,  MO Karlsson. (2013)
\emph{Are datasets for NLME models large enough for a bootstrap to provide reliable parameter uncertainty distributions?}
PAGE 2013.
\url{https://www.page-meeting.org/?abstract=2899}
}
\author{
Vipul Mann,  Matthew L. Fidler
}
